+++ packaging/ios/Subsurface-mobile.pro
@@ -163,6 +163,7 @@ SOURCES += ../../subsurface-mobile-main.cpp \
 	../../profile-widget/divepixmapitem.cpp \
 	../../profile-widget/divetooltipitem.cpp \
 	../../profile-widget/tankitem.cpp \
+	../../profile-widget/divehandler.cpp \
 	../../profile-widget/divelineitem.cpp \
 	../../profile-widget/diverectitem.cpp \
 	../../profile-widget/divetextitem.cpp
@@ -322,6 +323,7 @@ HEADERS += \
 	../../profile-widget/tankitem.h \
 	../../profile-widget/animationfunctions.h \
 	../../profile-widget/divecartesianaxis.h \
+	../../profile-widget/divehandler.h \
 	../../profile-widget/divelineitem.h \
 	../../profile-widget/divepixmapitem.h \
 	../../profile-widget/diverectitem.h \
